Update
Revisited this old project (April 2017) and cleaned up the programs for readability and simplicity (June 2018). Renamed the course project to reflect it as my own project.
Debugged the programs, mostly on missing data imputation and data merging.
Added read.csv() functions that loads the train and test datasets for the Jaccard metric.
Removed .txt files of the algorithms and replaced them with .RData files and added .Rhistory files for anyone wanting to look at the script (commands used).
The .txt files were messy and couldn't be loaded onto R -- it was only used for reading the code and outputs (which was easier for the professor to grade).
jpeg() functions output the graphs to the 'output' directory.
Added README.md.

Also note that at the time of this commit, Github linguist doesn't display R language symbol under the repository -- this is because Github's languages.yml does not support .RData and .Rhistory yet.